requirements.txt
opencv-python>=4.7.0
mediapipe>=0.10.0
numpy>=1.26.0
Pillow>=10.0.0
imutils>=0.5.4
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
eventlet>=0.35.0